Make variable content scrollable
The FormApplications' heights were set to auto, which meant when there
were too many trade goods or cities, the application would expand to the
entire screen before auto scaling.

To ensure buttons and headers remain visible, as well as let the trading
menu auto expand for results, the tables and lists were given a max
height and made scrollable.